refactor(native): dedupe native e2e task-polling helpers into helpers.ts
sandbox-resolution.e2e.test.ts and sandbox-restart-resurrection.e2e.test.ts
each defined byte-identical TaskSummary/listTasks/waitForFreshRunningTask
helpers. Move them into the shared helpers.ts, alongside the rest of this
suite's shared harness code.
