+++ apps/native/e2e/helpers.ts
@@ -20,7 +20,7 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 
 import { sleep } from "@decocms/shared/std";
-import { describe as bunDescribe } from "bun:test";
+import { describe as bunDescribe, expect } from "bun:test";
 
 /** Per-test-run bearer. 32+ chars, matches the daemon suite's convention
  *  (`daemon.e2e.helpers.ts`'s `DAEMON_TOKEN`) so a Rust binary honoring the
@@ -371,6 +371,56 @@ export async function stopLocalApi(
   }
   if (workdir && !opts.keepWorkdir)
     rmSync(workdir, { recursive: true, force: true });
+}
+
+// --- sandbox tasks -----------------------------------------------------------
+
+export interface TaskSummary {
+  id: string;
+  command: string;
+  status: string;
+  logName: string | null;
+}
+
+/** `GET /_sandbox/tasks` scoped to `handle` via the header (or the
+ * process-global registry when `handle` is omitted). */
+export async function listTasks(
+  a: LocalApi,
+  handle?: string,
+): Promise<TaskSummary[]> {
+  const res = await fetch(url(a, "/_sandbox/tasks"), {
+    headers: authHeaders(handle ? { "x-decocms-sandbox-handle": handle } : {}),
+  });
+  expect(res.status).toBe(200);
+  const body = (await res.json()) as { tasks: TaskSummary[] };
+  return body.tasks;
+}
+
+/** Polls `listTasks(a, handle)` until a RUNNING task named `logName` with an
+ * id different from every id in `excludeIds` shows up — proves a fresh
+ * dev-server task was spawned for THAT specific handle. */
+export async function waitForFreshRunningTask(
+  a: LocalApi,
+  handle: string,
+  logName: string,
+  excludeIds: Set<string>,
+  deadlineMs = 20_000,
+): Promise<TaskSummary> {
+  const deadline = Date.now() + deadlineMs;
+  while (Date.now() < deadline) {
+    const tasks = await listTasks(a, handle);
+    const fresh = tasks.find(
+      (t) =>
+        t.logName === logName &&
+        t.status === "running" &&
+        !excludeIds.has(t.id),
+    );
+    if (fresh) return fresh;
+    await sleep(200);
+  }
+  throw new Error(
+    `no fresh running "${logName}" task appeared for handle ${handle} within ${deadlineMs}ms`,
+  );
 }
 
 // --- SSE ---------------------------------------------------------------------
